refactor(frontend): move report logic in report folder
