+++ apps/frontend/components/Report/index.ts
@@ -1,0 +1,1 @@
+export { Report } from './Report';
